Name and skill client fix
Changed around the naming so its simpler and follows best practices. Also changed the skill page so it no longer use's use Client.
