+++ Portfolio/src/views/skills/gallery/gallery.css
@@ -38,6 +38,13 @@
   animation: forwards fadeUpEnter75-0 1.5s ease-in-out;
 }
 
+#gallery-link {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-color: transparent;
+}
+
 .gallery-container img,
 .gallery-container video {
   width: 100%;
