+++ Portfolio/src/views/skills/main/skill.css
@@ -178,6 +178,13 @@
   animation: forwards fadeUpEnter75-0 1s ease-in-out;
 }
 
+#gallery-link {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  background-color: transparent;
+}
+
 .gallery-link {
   border: none;
   height: 45px;
@@ -200,7 +207,6 @@
 }
 
 #Sep1 {
-  cursor: pointer;
   animation: fadeUpEnter75-0 0.7s forwards ease-in-out;
 }
