+++ Portfolio/src/views/skills/main/skill.jsx
@@ -1,8 +1,4 @@
-"use client";
 import "./skill.css";
-import { useRef, useEffect } from "react";
-//Text
-import { createAnimatable, utils } from "animejs";
 import {
   TechStacks,
   Header,
@@ -18,108 +14,22 @@ const client1 = "/skill/client1.webp";
 const dynamicAnimations = "/DynamicShot.webp";
 const fitShot = "/skill/FitShot.webp";
 const todoShot = "/skill/todoShot.webp";
-import { AddMember, RemoveMember } from "../../../utils/aniFrame.jsx";
 import {
   CreateFolder,
   CreateFeatured,
 } from "../../../components/skillPage/index.js";
 import Poly from "../../../components/svg/poly.jsx";
-import { useRouter } from "next/navigation";
+import PillAnimation from "./pillAnimation";
 
 function Skill() {
-  const root = useRef(null); // for animejs
-  const router = useRouter();
-  //Creates the lottie animation
-  //Manipulates the pills on mouse move
-  function onMouseMove(event, animatablePills, bluePills) {
-    const { clientX, clientY } = event;
-    bluePills.forEach((pill, index) => {
-      const bounding = pill.getBoundingClientRect();
-      const { left, top, width, height } = bounding;
-      // Calculate distance from the mouse to the center of the pill
-      const dx = clientX - (left + width / 2);
-      const dy = clientY - (top + height / 2);
-      const distance = Math.sqrt(dx * dx + dy * dy);
-
-      // Activate on distance
-      if (distance < 150) {
-        const hw = width / 2;
-        const hh = height / 2;
-        const x = utils.clamp(-(clientX - left - hw), -hw, hw);
-        const y = utils.clamp(-(clientY - top - hh), -hh, hh);
-        //Update x and y
-        animatablePills[index].x(x);
-        animatablePills[index].y(y);
-        pill.classList.add("hovered");
-      } else {
-        animatablePills[index].x(0);
-        animatablePills[index].y(0);
-        pill.classList.remove("hovered");
-      }
-    });
-  }
-  //Inits the animatable and return them
-  function initalize() {
-    // Selects them all
-    const bluePills = Array.from(
-      document.querySelectorAll(
-        ".blue-pill, .blue-pill h2, .blue-pill-2, .blue-pill-2 h2"
-      )
-    );
-    // Create individual animations for each BluePill
-    const animatablePills = bluePills.map((pill) =>
-      createAnimatable(pill, {
-        x: 400,
-        y: 400,
-        ease: "ease(3)",
-      })
-    );
-    return { animatablePills, bluePills };
-  }
-
-  useEffect(() => {
-    const InitVals = initalize();
-    let mouseX = 0;
-    let mouseY = 0;
-    let prevMouseX = null;
-    let prevMouseY = null;
-
-    const handleMouseMove = (event) => {
-      mouseX = event.clientX;
-      mouseY = event.clientY;
-    };
-
-    function update() {
-      // Only update if the mouse position has changed
-      if (mouseX !== prevMouseX || mouseY !== prevMouseY) {
-        onMouseMove(
-          { clientX: mouseX, clientY: mouseY },
-          InitVals.animatablePills,
-          InitVals.bluePills
-        );
-        prevMouseX = mouseX;
-        prevMouseY = mouseY;
-      }
-    }
-
-    AddMember(update);
-    window.addEventListener("mousemove", handleMouseMove);
-
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove);
-      RemoveMember(update);
-    };
-  }, []);
   return (
-    <div className="main-skill-container" ref={root}>
+    <div className="main-skill-container">
+      <PillAnimation />
       {/* For SEO */}
       <h1 style={{ display: "none" }}>Skills & Projects Portfolio</h1>
       {/* Over head container for tech stacks */}
-      <div
-        className="separator"
-        id="Sep1"
-        onClick={() => router.push("/skills/gallery")}
-      >
+      <div className="separator" id="Sep1">
+        <a id="gallery-link" href="/skills/gallery"></a>
         <h3>01.</h3>
         <h2>Known tech</h2>
         <hr></hr>
